+++ ElvUI_MerathilisUI/Core/Changelog.lua
@@ -12,11 +12,8 @@ local DISABLED_FONT_COLOR = DISABLED_FONT_COLOR
 
 local ChangeLogData = {
 	"Changes:",
-		"• !!IMPORTANT!!: With this version i will only support Retail",
-		"• Changed the file structure from my Plugin",
-		"• Added a time delay option to prevent posting double Announcements",
-		"• Removed the Reminder Module",
-		"• Added more Runes/more openable items to the AutoButtons",
+		"• Fix a ChatText error",
+		"• Take account for two new check boxes on the RaidManager",
 		-- "• ''",
 
 		" ",