+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### Changes ###
 
-  * FOR 10.1.7 >> ElvUI 13.40
-  * Only take account to specific Patch changes
+  * Fix a ChatText error
+  * Take account for two new check boxes on the RaidManager